fix: Parse legacy codes to the end
There was a buggy behavior where pressing `<S-F5>` (or other legacy
codes) would report both `<S-F5>` and `~`. This was happening because
the branching statement to go to parsing legacy CSI codes would check if
the buffer ended with ~ OR if the first value of the buffer after CSI
was '1'.

This second assumption would cause early processing of the keycode.
Because legacy key codes will all terminate with specific values
(`~ABCDEFHPQS`), we can check the final value in the buffer to make
sure membership is correct. Then we are good.

fixes: #13