feat: Add ReaderEvent for internal events
Crossterm-rs does this, and at first i thought it was uneccessary,
but after getting to the point where we need to be able to communicate
with the terminal and request KKP progressive enhancement status,
I realized that rather than doing all the parsing logic again,
it would make sense to do what crossterm does. Handle the parsing
all in the EventReader, and then only emit the KeyEvents.

The limiting/filtering of events is yet to be done, but this will
allow other kinds of events to be measured as well.